refactor(agent): max_tokens 默认不限制输出上限并统一类型
- 移除内置模板与 providers.toml 兜底中硬编码的 max_tokens 默认值，
  默认省略该字段交由服务端决定输出上限
- max_tokens 全链路类型统一为 u32，去掉 usize 中间态与多余的 as 强转
- 上下文预算字段 TokenBudget.max_tokens 重命名为 max_context_tokens，
  避免与输出上限混淆

+++ crates/model-provider/src/providers/deepseek.rs
@@ -108,7 +108,7 @@ struct DeepSeekRequest<'a> {
     #[serde(skip_serializing_if = "Option::is_none")]
     temperature: Option<f64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    max_tokens: Option<u64>,
+    max_tokens: Option<u32>,
     #[serde(skip_serializing_if = "Option::is_none")]
     stream: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
